requirements.full.txt
psycopg2-binary
rich
sqlalchemy
requests
fuzzywuzzy
python-levenshtein
pandas
python-dotenv
openpyxl
sentencepiece
art
flask
flask_restful
flask_cors
flasgger
web3
gunicorn
matplotlib

transformers
scikit-learn
torch
keybert